Add BsplineAllocator.
