+++ src/CMakeLists.txt
@@ -137,8 +137,6 @@ ENDIF()
     Utilities/XMLWriter.cpp
     Utilities/tinyxml/tinyxml2.cpp
     Utilities/qmcpack_version.cpp
-    Numerics/Spline2/einspline_allocator.cpp
-    Numerics/Spline2/bspline_allocator.cpp
     Numerics/Spline2/MultiBsplineData.cpp
     ${GITREV_TMP}
     )
