+++ src/Numerics/Spline2/MultiBsplineRef.hpp
@@ -21,7 +21,6 @@
 #define QMCPLUSPLUS_MULTIEINSPLINE_REF_HPP
 #include "config.h"
 #include <iostream>
-#include <Numerics/Spline2/bspline_allocator.hpp>
 #include <Numerics/Spline2/MultiBsplineData.hpp>
 #include <Numerics/Spline2/MultiBsplineEvalHelper.hpp>
 #include <stdlib.h>
